requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
joblib==0.14.1
lxml==4.5.0
nltk==3.5
pdfminer==20191125
Pillow==7.1.1
psycopg2==2.8.5
pycryptodome==3.9.7
python-docx==0.8.10
pytz==2019.3
regex==2020.4.4
requests==2.23.0
sqlparse==0.3.1
textblob==0.15.3
tqdm==4.45.0
urllib3==1.25.9
whitenoise==5.0.1
